Config-driven QA and analysis pipeline for geospatial datasets/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-audit-flow"
version = "0.1.0"
description = "A config-driven QA, cleaning, lightweight spatial analysis, and report-generation pipeline for geospatial datasets."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "geo-audit-flow contributors" }
]
keywords = ["gis", "geopandas", "geospatial", "data-quality", "cli", "qa"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
  "click>=8.1",
  "folium>=0.15",
  "geopandas>=0.14",
  "jinja2>=3.1",
  "pandas>=2.0",
  "pyogrio>=0.7",
  "pyyaml>=6.0",
  "shapely>=2.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.8",
  "pytest>=8.0",
  "ruff>=0.5",
  "types-PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/your-org/geo-audit-flow"
Issues = "https://github.com/your-org/geo-audit-flow/issues"
Documentation = "https://github.com/your-org/geo-audit-flow/tree/main/docs"

[project.scripts]
geoaudit = "geoaudit.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_unused_ignores = true
no_implicit_optional = true
